move edit and add meter from modal into the tab content section
