+++ src/app/facility/utility-data/energy-consumption/energy-source/edit-meter/edit-meter.component.css
@@ -1,0 +1,3 @@
+button span {
+    color: white;
+}
