+++ src/app/facility/utility-data/energy-consumption/energy-source/edit-meter/edit-meter.component.html
@@ -1,9 +1,14 @@
-<h4>Edit Meter</h4>
-<br>
-
-<app-edit-meter-form [(meterForm)]="meterForm" [meterDataExists]="meterDataExists"></app-edit-meter-form>
-
-<div class="saveCancel item-right">
-    <button class="btn btn-secondary" (click)="cancel()">Cancel</button>
-    <button class="btn btn-primary" (click)="saveChanges()" [disabled]="meterForm.invalid">Save</button>
+<div class="row">
+    <div class="col-12">
+        <div class="saveCancel item-right">
+            <button class="btn btn-secondary" (click)="cancel()">Cancel</button>
+            <button class="btn btn-primary" (click)="saveChanges()" [disabled]="meterForm.invalid">                
+                <span *ngIf="addOrEdit == 'add'">Add Meter</span>
+                <span *ngIf="addOrEdit == 'edit'">Save Changes</span>
+            </button>
+        </div>
+    </div>
+    <div class="col-12">
+        <app-edit-meter-form [(meterForm)]="meterForm" [meterDataExists]="meterDataExists"></app-edit-meter-form>
+    </div>
 </div>
